todo rehash

+++ src/test/bitset_test.cc
@@ -23,7 +23,7 @@ int main() {
 // bitmap: 0  0   1     1    0      1     0     1  |  0      1    0     1     1     1     0     1
                 0x11, 0x22, 0x33, 0x44, 0x55, 0x66, 0x77, 0x88, 0x99, 0xaa, 0xbb, 0xcc, 0xdd, 0xee};
 // index:  0  1   2     3    4      5     6     7     8     9     10    11    12    13    14    15        
-        turbo::unordered_map<int, int>::CellMeta256V2 meta256v2((char*)bitset_buffer256v2);
+        turbo::unordered_map<int, double>::CellMeta256V2 meta256v2((char*)bitset_buffer256v2);
         for (int i : meta256v2.EmptyBitSet()) {
             printf("empty slot %d\n", i);
         }
